feat: SAA + Navigation 구현
