+++ app/src/main/AndroidManifest.xml
@@ -34,14 +34,6 @@
 
         </activity>
         <activity
-            android:name=".ui.auth.AuthActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.Swapit">
-
-        </activity>
-
-        <activity
             android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
             android:exported="true">
             <intent-filter>
