+++ app/src/main/java/com/example/swapit/MainActivity.kt
@@ -3,7 +3,10 @@ package com.example.swapit
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
+import com.example.swapit.domain.repository.SocialLoginRepository
+import com.example.swapit.ui.auth.LoginViewModel
 import com.example.swapit.ui.navigation.NavigationModule
 
 class MainActivity : ComponentActivity() {
@@ -12,7 +15,13 @@ class MainActivity : ComponentActivity() {
         setContent {
             val navController = rememberNavController()
             val navigationModule = NavigationModule()
-            navigationModule.NavigationGraph(navController)
+            val loginViewModel: LoginViewModel = viewModel(
+                factory = LoginViewModel.factory(
+                    application,
+                    SocialLoginRepository.instance(this)
+                )
+            )
+            navigationModule.NavigationGraph(navController,loginViewModel)
         }
     }
 }
