+++ app/src/main/java/com/example/swapit/ui/navigation/NavigationModule.kt
@@ -1,11 +1,15 @@
 package com.example.swapit.ui.navigation
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import com.example.swapit.ui.alert.AlertScreen
+import com.example.swapit.ui.auth.LoginScreen
+import com.example.swapit.ui.auth.LoginViewModel
 import com.example.swapit.ui.chat.ChatListScreen
 import com.example.swapit.ui.post.PostProductScreen
 import com.example.swapit.ui.post.PostProductViewModel
@@ -19,11 +23,25 @@ import com.example.swapit.ui.user.UserInfoScreen
 
 class NavigationModule {
     @Composable
-    fun NavigationGraph(navController: NavHostController) {
+    fun NavigationGraph(
+        navController: NavHostController,
+        loginViewModel: LoginViewModel,
+    ) {
+        val isLoggedIn by loginViewModel.isLoggedIn.collectAsState()
+        val startDestination =
+            if (isLoggedIn) BottomNavItem.Shopping.screenRoute else BottomNavItem.Login.screenRoute
+
         NavHost(
             navController = navController,
-            startDestination = BottomNavItem.Shopping.screenRoute,
+            startDestination = startDestination,
         ) {
+            composable(BottomNavItem.Login.screenRoute) {
+                LoginScreen(
+                    navController = navController,
+                    viewModel = loginViewModel,
+                )
+            }
+
             composable(BottomNavItem.Shopping.screenRoute) {
                 ShoppingScreen(navController)
             }
